Tests/LibWeb/Text/expected/HTML/parser-streams-style-split.txt

rgb(1, 2, 3)
styled
